#164: create smartwrap abstraction
