+++ src/main/kotlin/utils/LoggingSmartWrap.kt
@@ -1,0 +1,8 @@
+package utils
+
+class LoggingSmartWrap : SmartWrap {
+    override fun invoke(input: String): String {
+        println(input)
+        return input
+    }
+}
