+++ src/main/kotlin/utils/SmartWrapDependencyInjection.kt
@@ -1,0 +1,9 @@
+package utils
+
+import org.koin.dsl.factory
+import org.koin.dsl.module
+
+val smartWrapModule =
+    module {
+        factory<SmartWrap> { LoggingSmartWrap() }
+    }
